easy-deploy create-project now more idempotent
The create-project function of the easy-deploy script is now more
idempotent; it does not error out if things already exist, and only
creates symlinks if they do not exist. It does touch the text files
within each project (samples-assembly.txt, etc.). An environment
variable, OMIT_UGER_PROJECT_FILES is now checked, and the run-pipe.sh
scripts for UGER are only symlinked in if the env var is not set. These
changes support use with the Docker image and dsub.